Day 2 solutions uploaded
I won't lie, this was tough. Mainly because I was an idiot and realized that it took way less time to run than I thought it would, so I didn't have to do a complicated algorithm approach for part 1 (I could have just done a direct iteration). Maybe one day I will go back and rework it.

After trying to do some weird thing for part 2, I came across this AMAZING comment on Stack Overflow, which became the basis from my solution: https://stackoverflow.com/questions/55823298/how-do-i-check-if-a-string-is-entirely-made-of-the-same-substring/55840779#55840779

HUGE thanks to the people who mentioned the theorem there! Part of my solution for Part 2 is extremely lazy (the part where I just added a separate case for len = 2 bc I was too lazy to debug it lol), but it got the job done :)