Day 3 solutions uploaded
Day 3! First part was pretty easy, but the second one definitely took some more thought. In case you're wondering where the number 11 came from, it is 12 minus the digit that was selected to be added to the maximum string, because we need the complete string to have at least 12 digits. The logic is weird, and it took a bit to figure out, but I did it :D

For a better visual explanation, see the image I put in the folder.